Move MethodDefinition logic to generate IL into extension.
