+++ src/dotnet-ildasm.Tests/Infrastructure/InstructionExtensionsShould.cs
@@ -21,28 +21,6 @@ namespace DotNet.Ildasm.Tests.Infrastructure
             _assemblyDefinition = DataHelper.SampleAssembly.Value;
         }
 
-        [Theory]
-        [InlineData("PublicClass", "PublicVoidMethod", ".method public hidebysig instance void PublicVoidMethod() cil managed")]
-        [InlineData("PublicClass", "PublicVoidMethodSingleParameter", ".method public hidebysig instance void PublicVoidMethodSingleParameter(string parameter1) cil managed")]
-        [InlineData("PublicClass", "PublicVoidMethodTwoParameters", ".method public hidebysig instance void PublicVoidMethodTwoParameters(string parameter1, int32 parameter2) cil managed")]
-        [InlineData("PublicClass", "PublicVoidMethodParams", ".method public hidebysig instance void PublicVoidMethodParams(string[] parameters) cil managed")]
-        [InlineData("PublicClass", "set_Property1", ".method public hidebysig specialname instance void set_Property1(string 'value') cil managed")]
-        [InlineData("PublicAbstractClass", "PublicAbstractMethod", ".method public hidebysig newslot abstract virtual instance void PublicAbstractMethod() cil managed")]
-        [InlineData("PublicAbstractClass", "PublicImplementedMethod", ".method public hidebysig instance void PublicImplementedMethod() cil managed")]
-        [InlineData("DerivedPublicClass", "PublicAbstractMethod", ".method public hidebysig virtual instance void PublicAbstractMethod() cil managed")]
-        [InlineData("DerivedPublicClass", "PublicAbstractSealedMethod", ".method public hidebysig virtual final instance void PublicAbstractSealedMethod() cil managed")]
-        [InlineData("DerivedPublicClass", "PublicImplementedMethod", ".method public hidebysig instance void PublicImplementedMethod() cil managed")]
-        public void Write_Method_Signature(string className, string methodName, string expectedIL)
-        {
-            var type = _assemblyDefinition.MainModule.Types.FirstOrDefault(x => x.Name == className);
-            var method = type.Methods.FirstOrDefault(x => x.Name == methodName);
-            var methodProcessor = new MethodProcessor(_outputWriterMock);
-
-            method.WriteILSignature(_outputWriter);
-
-            Assert.Equal(expectedIL, _outputWriter.ToString());
-        }
-
         [Fact]
         public void Be_Able_To_Write_Literal_String()
         {
