+++ src/dotnet-ildasm/InstructionExtensions.cs
@@ -72,76 +72,5 @@ namespace DotNet.Ildasm
 
             writer.Write(")");
         }
-
-
-
-        public static void WriteILSignature(this MethodDefinition method, IOutputWriter writer)
-        {
-            writer.Write(".method");
-
-            if (method.IsPublic)
-                writer.Write(" public");
-            else if (method.IsPrivate)
-                writer.Write(" private");
-
-            if (method.IsHideBySig)
-                writer.Write(" hidebysig");
-
-            if (method.IsNewSlot)
-                writer.Write(" newslot");
-
-            if (method.IsAbstract)
-                writer.Write(" abstract");
-
-            if (method.IsVirtual)
-                writer.Write(" virtual");
-
-            if (method.IsSpecialName)
-                writer.Write(" specialname");
-
-            if (method.IsRuntimeSpecialName)
-                writer.Write(" rtspecialname");
-
-            if (method.IsFinal)
-                writer.Write(" final");
-
-            if (!method.IsStatic)
-                writer.Write(" instance");
-            else
-                writer.Write(" static");
-
-            writer.Write($" {method.ReturnType.ToIL()}");
-            writer.Write($" {method.Name}");
-
-            WriteMethodSignatureParameters(method, writer);
-
-            if (method.IsManaged)
-                writer.Write(" cil managed");
-        }
-
-        private static void WriteMethodSignatureParameters(MethodDefinition method, IOutputWriter writer)
-        {
-            writer.Write("(");
-
-            if (method.HasParameters)
-            {
-                for (int i = 0; i < method.Parameters.Count; i++)
-                {
-                    if (i > 0)
-                        writer.Write(", ");
-
-                    var parameterDefinition = method.Parameters[i];
-                    writer.Write($"{parameterDefinition.ParameterType.ToIL()} ");
-
-                    if (parameterDefinition.Name == "value")
-                        writer.Write($"'{parameterDefinition.Name}'");
-                    else
-                        writer.Write(parameterDefinition.Name);
-                }
-            }
-
-            writer.Write(")");
-        }
-
     }
 }
